+++ src/lib/connection/XMPPServiceHandler/RBVoiceEventHandler.ts
@@ -199,6 +199,9 @@ class RBVoiceEventHandler extends GenericHandler {
                     case "roomlobby":
                         //treated in conversationEventHandler::onRoomLobbyMessageReceived(node);
                         break;
+                    case "newIncoming":
+                        //treated in conversationEventHandler::onNewIncomingMessageReceived(node);
+                        break;
                     case "file":
                         break;
                     case "thumbnail":