+++ src/lib/connection/XMPPServiceHandler/TasksEventHandler.ts
@@ -168,6 +168,9 @@ class TasksEventHandler extends GenericHandler {
                     case "roomlobby":
                         //treated in conversationEventHandler::onRoomLobbyMessageReceived(node);
                         break;
+                    case "newIncoming":
+                        //treated in conversationEventHandler::onNewIncomingMessageReceived(node);
+                        break;
                     case "file":
                         // treated in conversationEventHandler
                         break;