+++ src/lib/connection/XMPPServiceHandler/alertEventHandler.ts
@@ -373,6 +373,9 @@ class AlertEventHandler extends GenericHandler {
                     case "roomlobby":
                         //treated in conversationEventHandler::onRoomLobbyMessageReceived(node);
                         break;
+                    case "newIncoming":
+                        //treated in conversationEventHandler::onNewIncomingMessageReceived(node);
+                        break;
                     case "file":
                         break;
                     case "thumbnail":