+++ src/lib/connection/XMPPServiceHandler/invitationEventHandler.ts
@@ -93,6 +93,9 @@ class InvitationEventHandler extends GenericHandler {
                     case "roomlobby":
                         //treated in conversationEventHandler::onRoomLobbyMessageReceived(node);
                         break;
+                    case "newIncoming":
+                        //treated in conversationEventHandler::onNewIncomingMessageReceived(node);
+                        break;
                     case "file":
                         // treated in conversationEventHandler
                         break;