+++ src/lib/connection/XMPPServiceHandler/presenceEventHandler.ts
@@ -392,6 +392,9 @@ class PresenceEventHandler extends GenericHandler {
                     case "roomlobby":
                         //treated in conversationEventHandler::onRoomLobbyMessageReceived(node);
                         break;
+                    case "newIncoming":
+                        //treated in conversationEventHandler::onNewIncomingMessageReceived(node);
+                        break;
                     case "file":
                         // treated in conversationEventHandler
                         break;